Add unit tests for get_notice_text_from_key_files #136
