+++ dejacode_toolkit/scancodeio.py
@@ -449,7 +449,7 @@ def get_package_download_url(project_data):
         return input_sources[0].get("download_url", None)
 
 
-def get_notice_text_from_key_files(scan_summary, separator="\n---\n\n"):
+def get_notice_text_from_key_files(scan_summary, separator="\n\n---\n\n"):
     """
     Return a generate notice_text from the key files contained in the provided
     ``scan_summary``.
@@ -460,5 +460,7 @@ def get_notice_text_from_key_files(scan_summary, separator="\n---\n\n"):
     # of a `is_notice` attribute.
     notice_files = [key_file for key_file in key_files if "notice" in key_file.get("name").lower()]
 
-    notice_text = separator.join([notice_file.get("content") for notice_file in notice_files])
+    notice_text = separator.join(
+        [notice_file.get("content").strip() for notice_file in notice_files]
+    )
     return notice_text
